reorg of scripts and semi-split of codalib/codadaemon
