+++ scripts/status.py
@@ -1,4 +1,6 @@
 #! /usr/bin/env python3
 import codalib
+import json
 if __name__ == "__main__":
-    codalib.cluster_status()
+    result = codalib.filtered_status()
+    print(json.dumps(result, sort_keys=True, indent=4))
