+++ scripts/utilities/block_wait
